letsencrypt: improve renew scripts
* renewAll now calls restore.sh in case it can't find the acme.sh directory
* renewAll now checks for changes before writing a dockercloud.key file
* updateDockercloud now deletes the dockercloud.key file on success

A cron job can now be established that does 'renewAll.sh && updateDockercloud.sh && backup.sh1

+++ letsencrypt/scripts/renewAll.sh
@@ -1,16 +1,38 @@
 #! /bin/bash
 
+[[ -z "$S3_BUCKET" ]] && echo "s3 bucket not set, exiting" && exit -1;
+
+[[ ! -d ~/.acme.sh/ ]] && echo "restoring from s3" && ./restore.sh
+
+# find all directories which correspond to domains & list them so we can see what changed
+DOMAINS=`find ~/.acme.sh/ -maxdepth 1 -mindepth 1 -type d | grep '\.[^/]\{2,\}$'`
+if [[ -z "$DOMAINS" ]]; then
+    echo "restoring from s3"
+    ./restore.sh
+
+    DOMAINS=`find ~/.acme.sh/ -maxdepth 1 -mindepth 1 -type d | grep '\.[^/]\{2,\}$'`
+    [[ -z "$DOMAINS" ]] && echo "error restoring from bucket $S3_BUCKET" && exit -1;
+fi
+
 set -e
 
-[[ -z "$S3_BUCKET" ]] && echo "s3 bucket not set, exiting" && exit -1;
+declare -A stats
+while read -r path; do
+    domain=`basename $path`
+
+    stats[$domain]=`ls -l $path`
+done <<< "$DOMAINS"
 
 acme.sh --renewAll
 
-# find all directories which correspond to domains
+# find all domains again and compare
 DOMAINS=`find ~/.acme.sh/ -maxdepth 1 -mindepth 1 -type d | grep '\.[^/]\{2,\}$'`
 
 while read -r path; do
     domain=`basename $path`
+
+    # nothing changed, skip creating the dockercloud file
+    [[ "${stats[$domain]}" == "`ls -l $path`" ]] && continue;
 
     # combine the certificate with the private key
     cat $path/fullchain.cer $path/$domain.key  > $path/dockercloud.key
